finished fine-tuning of file-format specific parameters, implemented Dynamic Minimum Degree Construction Heuristic (used for jpegs), if no -a argument is given, a file-format specific combination is used, speedup of getDegree() calculation, more unit tests, more system tests, fixed system test exit status, updated man page, removed PCM32 encoding for au files, bugfixes, some minor changes

+++ steghide/tests/DummyFile.cc
@@ -91,6 +91,6 @@ void DummyFile::createGraph (std::vector<std::list<UWORD16> >& adjlist, BitStrin
 	}
 
 	(*f) = (CvrStgFile*) new DummyFile (2 * numvertices, svam) ;
-	(*bs) = new BitString (numvertices) ; // zeros only
+	(*bs) = new BitString ((unsigned long) numvertices) ; // zeros only
 	(*s) = new Selector (2 * numvertices) ; // identity permutation
 }
